requirements.txt
javaobj-py3
pycryptodome
python-multipart
fastapi
uvicorn